fix all rv32 definitions and improve tests
