+++ examples/RV32F.toml
@@ -130,7 +130,7 @@ flw = { mask = 0x707f, match = 0x2007 }
 [store]
 type = "S"
 [store.repr]
-default = "$name$ %rdf%, %imm%(%rs1i%)"
+default = "$name$ %rs2f%, %imm%(%rs1i%)"
 [store.instructions]
 fsw = { mask = 0x707f, match = 0x2027 }
 
@@ -148,6 +148,7 @@ fnmsub_s = { mask = 0x600007f, match = 0x4b }
 type = "Rfff"
 [arith.repr]
 default = "$name$ %rdf%, %rs1f%, %rs2f%"
+fsqrt_s = "$name$ %rdf%, %rs1f%"
 [arith.instructions]
 fadd_s = { mask = 0xfe00007f, match = 0x53 }
 fsub_s = { mask = 0xfe00007f, match = 0x8000053 }
@@ -163,7 +164,7 @@ fsqrt_s = { mask = 0xfff0007f, match = 0x58000053 }
 [i2f]
 type = "Rfif"
 [i2f.repr]
-default = "$name$ %rdf% %rs1i%"
+default = "$name$ %rdf%, %rs1i%"
 [i2f.instructions]
 fcvt_s_w = { mask = 0xfff0007f, match = 0xd0000053 }
 fcvt_s_wu = { mask = 0xfff0007f, match = 0xd0100053 }
@@ -172,7 +173,7 @@ fmv_w_x = { mask = 0xfff0707f, match = 0xf0000053 }
 [f2i]
 type = "Rffi"
 [f2i.repr]
-default = "$name$ %rdi% %rs1f%"
+default = "$name$ %rdi%, %rs1f%"
 feq_s = "$name$ %rdi%, %rs1f%, %rs2f%"
 flt_s = "$name$ %rdi%, %rs1f%, %rs2f%"
 fle_s = "$name$ %rdi%, %rs1f%, %rs2f%"
