+++ examples/RV32I.toml
@@ -9,9 +9,11 @@ names = [
     "csri",
     "load",
     "store",
+    "shift",
     "branch",
     "jump",
     "upper",
+    "fence"
 ]
 parts = [
     [
@@ -50,14 +52,34 @@ parts = [
         "Register_int",
     ],
     [
+        "shamt",
+        5,
+        "u8"
+    ],
+    [
         "imm",
         32,
         "VInt",
     ],
+    [
+        "pred",
+        4,
+        "Register_fence"
+    ],
+    [
+        "succ",
+        4,
+        "Register_fence"
+    ],
+    [
+        "discard",
+        32,
+        "VInt"
+    ]
 ]
 
 [type]
-names = ["R", "I", "I_shift", "S", "B", "U", "J", "CSR", "CSRI"]
+names = ["R", "I", "I_shift", "S", "B", "U", "J", "CSR", "CSRI", "Fence"]
 R = [
     { name = "funct7", top = 6, bot = 0 },
     { name = "rs2", top = 4, bot = 0 },
@@ -74,8 +96,8 @@ I = [
     { name = "opcode", top = 6, bot = 0 },
 ]
 I_shift = [
-    { name = "shtyp", top = 6, bot = 0 },
-    { name = "imm", top = 4, bot = 0 },
+    { name = "funct7", top = 6, bot = 0 },
+    { name = "shamt", top = 4, bot = 0 },
     { name = "rs1", top = 4, bot = 0 },
     { name = "funct3", top = 2, bot = 0 },
     { name = "rd", top = 4, bot = 0 },
@@ -127,6 +149,15 @@ J = [
     { name = "rd", top = 4, bot = 0 },
     { name = "opcode", top = 6, bot = 0 },
 ]
+Fence = [
+    { name = "discard", top = 3, bot = 0 },
+    { name = "pred", top = 3, bot = 0 },
+    { name = "succ", top = 3, bot = 0 },
+    { name = "rs1", top = 4, bot = 0 },
+    { name = "funct3", top = 2, bot = 0 },
+    { name = "rd", top = 4, bot = 0 },
+    { name = "opcode", top = 6, bot = 0 },
+]
 
 [reg]
 type = "R"
@@ -150,6 +181,8 @@ type = "I"
 [imm.repr]
 default = "$name$ %rd%, %rs1%, %imm%"
 jalr = "$name$ %rd%, %imm%(%rs1%)"
+slti = "$name$ %rd%, %rs1%, %imm%"
+sltiu = "$name$ %rd%, %rs1%, %imm%"
 ecall = "$name$"
 ebreak = "$name$"
 [imm.instructions]
@@ -158,6 +191,8 @@ xori = { mask = 0x707f, match = 0x4013 }
 ori = { mask = 0x707f, match = 0x6013 }
 andi = { mask = 0x707f, match = 0x7013 }
 jalr = { mask = 0x707f, match = 0x67 }
+slti = { mask = 0x707f, match = 0x2013 }
+sltiu = { mask = 0x707f, match = 0x3013 }
 ecall = { mask = 0xffffffff, match = 0x73 }
 ebreak = { mask = 0xffffffff, match = 0x100073 }
 
@@ -182,13 +217,11 @@ csrrwi = { mask = 0x707f, match = 0x5073, unsigned = true }
 [shift]
 type = "I_shift"
 [shift.repr]
-default = "$name$ %rd%, %rs1%, %imm%"
+default = "$name$ %rd%, %rs1%, %shamt%"
 [shift.instructions]
-slli = { mask = 0xfe00707f, match = 0x1013 }
-srli = { mask = 0xfe00707f, match = 0x5013 }
-srai = { mask = 0xfe00707f, match = 0x40005013 }
-slti = { mask = 0x707f, match = 0x2013 }
-sltiu = { mask = 0x707f, match = 0x3013 }
+slli = { mask = 0xfe00707f, match = 0x1013}
+srai = { mask = 0xfe00707f, match = 0x40005013}
+srli = { mask = 0xfe00707f, match = 0x5013}
 
 [load]
 type = "I"
@@ -237,8 +270,15 @@ default = "$name$ %rd%, %imm%"
 lui = { mask = 0x7f, match = 0x37, unsigned = true }
 auipc = { mask = 0x7f, match = 0x17, unsigned = true  }
 
+[fence]
+type = "Fence"
+[fence.repr]
+default = "fence %pred%, %succ%"
+[fence.instructions]
+fence = { mask = 0x707f, match = 0xf}
+
 [register]
-names = ["Register_int", "Register_csr"]
+names = ["Register_int", "Register_csr", "Register_fence"]
 number = 32
 Register_int = [
     "zero",
@@ -599,3 +639,148 @@ Register_int = [
 0xF14 = "mhartid"
 0xF15 = "mconfigptr"
 0xFB0 = "mtopi"
+0x114 = "sieh"
+0x154 = "siph"
+0x15D = "stimecmph"
+0x214 = "vsieh"
+0x254 = "vsiph"
+0x25D = "vstimecmph"
+0x615 = "htimedeltah"
+0x613 = "hidelegh"
+0x618 = "hvienh"
+0x61A = "henvcfgh"
+0x655 = "hviph"
+0x656 = "hviprio1h"
+0x657 = "hviprio2h"
+0x61C = "hstateen0h"
+0x61D = "hstateen1h"
+0x61E = "hstateen2h"
+0x61F = "hstateen3h"
+0xC80 = "cycleh"
+0xC81 = "timeh"
+0xC82 = "instreth"
+0xC83 = "hpmcounter3h"
+0xC84 = "hpmcounter4h"
+0xC85 = "hpmcounter5h"
+0xC86 = "hpmcounter6h"
+0xC87 = "hpmcounter7h"
+0xC88 = "hpmcounter8h"
+0xC89 = "hpmcounter9h"
+0xC8A = "hpmcounter10h"
+0xC8B = "hpmcounter11h"
+0xC8C = "hpmcounter12h"
+0xC8D = "hpmcounter13h"
+0xC8E = "hpmcounter14h"
+0xC8F = "hpmcounter15h"
+0xC90 = "hpmcounter16h"
+0xC91 = "hpmcounter17h"
+0xC92 = "hpmcounter18h"
+0xC93 = "hpmcounter19h"
+0xC94 = "hpmcounter20h"
+0xC95 = "hpmcounter21h"
+0xC96 = "hpmcounter22h"
+0xC97 = "hpmcounter23h"
+0xC98 = "hpmcounter24h"
+0xC99 = "hpmcounter25h"
+0xC9A = "hpmcounter26h"
+0xC9B = "hpmcounter27h"
+0xC9C = "hpmcounter28h"
+0xC9D = "hpmcounter29h"
+0xC9E = "hpmcounter30h"
+0xC9F = "hpmcounter31h"
+0x310 = "mstatush"
+0x313 = "midelegh"
+0x314 = "mieh"
+0x318 = "mvienh"
+0x319 = "mviph"
+0x31A = "menvcfgh"
+0x31C = "mstateen0h"
+0x31D = "mstateen1h"
+0x31E = "mstateen2h"
+0x31F = "mstateen3h"
+0x354 = "miph"
+0x721 = "mcyclecfgh"
+0x722 = "minstretcfgh"
+0x723 = "mhpmevent3h"
+0x724 = "mhpmevent4h"
+0x725 = "mhpmevent5h"
+0x726 = "mhpmevent6h"
+0x727 = "mhpmevent7h"
+0x728 = "mhpmevent8h"
+0x729 = "mhpmevent9h"
+0x72A = "mhpmevent10h"
+0x72B = "mhpmevent11h"
+0x72C = "mhpmevent12h"
+0x72D = "mhpmevent13h"
+0x72E = "mhpmevent14h"
+0x72F = "mhpmevent15h"
+0x730 = "mhpmevent16h"
+0x731 = "mhpmevent17h"
+0x732 = "mhpmevent18h"
+0x733 = "mhpmevent19h"
+0x734 = "mhpmevent20h"
+0x735 = "mhpmevent21h"
+0x736 = "mhpmevent22h"
+0x737 = "mhpmevent23h"
+0x738 = "mhpmevent24h"
+0x739 = "mhpmevent25h"
+0x73A = "mhpmevent26h"
+0x73B = "mhpmevent27h"
+0x73C = "mhpmevent28h"
+0x73D = "mhpmevent29h"
+0x73E = "mhpmevent30h"
+0x73F = "mhpmevent31h"
+0x740 = "mnscratch"
+0x741 = "mnepc"
+0x742 = "mncause"
+0x744 = "mnstatus"
+0x757 = "mseccfgh"
+0xB80 = "mcycleh"
+0xB82 = "minstreth"
+0xB83 = "mhpmcounter3h"
+0xB84 = "mhpmcounter4h"
+0xB85 = "mhpmcounter5h"
+0xB86 = "mhpmcounter6h"
+0xB87 = "mhpmcounter7h"
+0xB88 = "mhpmcounter8h"
+0xB89 = "mhpmcounter9h"
+0xB8A = "mhpmcounter10h"
+0xB8B = "mhpmcounter11h"
+0xB8C = "mhpmcounter12h"
+0xB8D = "mhpmcounter13h"
+0xB8E = "mhpmcounter14h"
+0xB8F = "mhpmcounter15h"
+0xB90 = "mhpmcounter16h"
+0xB91 = "mhpmcounter17h"
+0xB92 = "mhpmcounter18h"
+0xB93 = "mhpmcounter19h"
+0xB94 = "mhpmcounter20h"
+0xB95 = "mhpmcounter21h"
+0xB96 = "mhpmcounter22h"
+0xB97 = "mhpmcounter23h"
+0xB98 = "mhpmcounter24h"
+0xB99 = "mhpmcounter25h"
+0xB9A = "mhpmcounter26h"
+0xB9B = "mhpmcounter27h"
+0xB9C = "mhpmcounter28h"
+0xB9D = "mhpmcounter29h"
+0xB9E = "mhpmcounter30h"
+0xB9F = "mhpmcounter31h"
+
+[register.Register_fence]
+0b0000 = "-"
+0b0001 = "w"
+0b0010 = "r"
+0b0011 = "rw"
+0b0100 = "o"
+0b0101 = "ow"
+0b0110 = "or"
+0b0111 = "orw"
+0b1000 = "i"
+0b1001 = "iw"
+0b1010 = "ir"
+0b1011 = "irw"
+0b1100 = "io"
+0b1101 = "iow"
+0b1110 = "ior"
+0b1111 = "iorw"
